Create Task 05
Write a program that reads two strings for playing the game of Rock-Paper-Scissors. If the strings entered by the user are not 'Paper', 'Rock' or 'Scissors', the program keeps on prompting the user to enter new values. If valid strings are inserted, repeat the loop, until one of the sides wins. (You can use alert instead of console.log).